Made QSStrings 64-bit clean.

+++ QSStrings.m
@@ -152,7 +152,7 @@ static const short _base64DecodingTable[256] = {
 	char * strResult;
 
 	// Get the Raw Data length and ensure we actually have data
-	int intLength = [objData length];
+	size_t intLength = [objData length];
 	if (intLength == 0) return nil;
 
 	// Setup the String-based Result placeholder and pointer within that placeholder
@@ -192,7 +192,7 @@ static const short _base64DecodingTable[256] = {
 + (NSData *)decodeBase64WithString:(NSString *)strBase64 {
 	const char * objPointer = [strBase64 cStringUsingEncoding:NSASCIIStringEncoding];
 	if (objPointer == NULL)  return nil;
-	int intLength = strlen(objPointer);
+	size_t intLength = strlen(objPointer);
 	int intCurrent;
 	int i = 0, j = 0, k;
